Add edit tests

+++ front/app/modules/commercial/insights/admin/containers/Insights/Edit/InputsTable/InputsTableRow.tsx
@@ -1,15 +1,30 @@
 import React from 'react';
-import useIdea from 'hooks/useIdea';
+import { withRouter, WithRouterProps } from 'react-router';
+
+// utils
 import { isNilOrError } from 'utils/helperUtils';
+
+// services
 import {
   IInsightsInputData,
   deleteInsightsInputCategory,
 } from 'modules/commercial/insights/services/insightsInputs';
-import { withRouter, WithRouterProps } from 'react-router';
+
+// hooks
+import useIdea from 'hooks/useIdea';
+
+// styles
+import styled from 'styled-components';
+
+// components
 import { Checkbox } from 'cl2-component-library';
 import T from 'components/T';
 import Tag from 'modules/commercial/insights/admin/components/Tag';
-import styled from 'styled-components';
+
+// intl
+import { injectIntl } from 'utils/cl-intl';
+import { InjectedIntlProps } from 'react-intl';
+import messages from '../../messages';
 
 const TagList = styled.div`
   > * {
@@ -20,20 +35,31 @@ const TagList = styled.div`
 
 type InputsTableRow = {
   input: IInsightsInputData;
-} & WithRouterProps;
+} & WithRouterProps &
+  InjectedIntlProps;
 
-const InputsTableRow = ({ input, params: { viewId } }: InputsTableRow) => {
+const InputsTableRow = ({
+  input,
+  params: { viewId },
+  intl: { formatMessage },
+}: InputsTableRow) => {
   const idea = useIdea({ ideaId: input.relationships?.source.data.id });
 
   if (isNilOrError(idea)) {
     return null;
   }
 
-  const handleRemoveCategory = (categoryId: string) => () =>
-    deleteInsightsInputCategory(viewId, input.id, categoryId);
+  const handleRemoveCategory = (categoryId: string) => () => {
+    const deleteMessage = formatMessage(
+      messages.inputsTableDeleteCategoryConfirmation
+    );
+    if (window.confirm(deleteMessage)) {
+      deleteInsightsInputCategory(viewId, input.id, categoryId);
+    }
+  };
 
   return (
-    <tr tabIndex={0}>
+    <tr tabIndex={0} data-testid="insightsInputsTableRow">
       <td>
         <Checkbox checked={false} onChange={() => {}} />
       </td>
@@ -57,4 +83,4 @@ const InputsTableRow = ({ input, params: { viewId } }: InputsTableRow) => {
   );
 };
 
-export default withRouter(InputsTableRow);
+export default injectIntl(withRouter(InputsTableRow));
